+++ tests/scalers/azure/azure_service_bus_topic_regex/azure_service_bus_topic_regex_test.go
@@ -1,7 +1,7 @@
 //go:build e2e
 // +build e2e
 
-package azure_service_bus_topic_test
+package azure_service_bus_topic_regex_test
 
 import (
 	"context"
@@ -21,21 +21,21 @@ import (
 )
 
 // Load environment variables from .env file
-var _ = godotenv.Load("../../.env")
+var _ = godotenv.Load("../../../.env")
 
 const (
-	testName = "azure-service-bus-topic-test"
+	testName = "azure-sb-topic-regex-test"
 )
 
 var (
-	connectionString = os.Getenv("AZURE_SERVICE_BUS_CONNECTION_STRING")
-	testNamespace    = fmt.Sprintf("%s-ns", testName)
-	secretName       = fmt.Sprintf("%s-secret", testName)
-	deploymentName   = fmt.Sprintf("%s-deployment", testName)
-	triggerAuthName  = fmt.Sprintf("%s-ta", testName)
-	scaledObjectName = fmt.Sprintf("%s-so", testName)
-	topicName        = fmt.Sprintf("%s-topic-%d", testName, GetRandomNumber())
-	subscriptionName = fmt.Sprintf("%s-subscription-%d", testName, GetRandomNumber())
+	connectionString   = os.Getenv("AZURE_SERVICE_BUS_CONNECTION_STRING")
+	testNamespace      = fmt.Sprintf("%s-ns", testName)
+	secretName         = fmt.Sprintf("%s-secret", testName)
+	deploymentName     = fmt.Sprintf("%s-deployment", testName)
+	triggerAuthName    = fmt.Sprintf("%s-ta", testName)
+	scaledObjectName   = fmt.Sprintf("%s-so", testName)
+	topicName          = fmt.Sprintf("%s-topic-%d", testName, GetRandomNumber())
+	subscriptionPrefix = fmt.Sprintf("%s-subscription-%d", testName, GetRandomNumber())
 )
 
 type templateData struct {
@@ -47,6 +47,7 @@ type templateData struct {
 	ScaledObjectName string
 	TopicName        string
 	SubscriptionName string
+	Operation        string
 }
 
 const (
@@ -109,15 +110,22 @@ spec:
   pollingInterval: 5
   cooldownPeriod: 10
   minReplicaCount: 0
-  maxReplicaCount: 1
+  maxReplicaCount: 10
   triggers:
-  - type: azure-servicebus
-    metadata:
-      topicName: {{.TopicName}}
-      subscriptionName: {{.SubscriptionName}}
-      activationMessageCount: "5"
-    authenticationRef:
-      name: {{.TriggerAuthName}}
+    - type: azure-servicebus
+      metadata:
+        topicName: {{.TopicName}}
+        subscriptionName: {{.SubscriptionName}}
+        messageCount: "1"
+        useRegex: "true"
+        operation: {{.Operation}}
+      authenticationRef:
+        name: {{.TriggerAuthName}}
+  advanced:
+    horizontalPodAutoscalerConfig:
+      behavior:
+        scaleDown:
+          stabilizationWindowSeconds: 5
 `
 )
 
@@ -137,13 +145,11 @@ func TestScaler(t *testing.T) {
 		"replica count should be 0 after 1 minute")
 
 	// test scaling
-	testActivation(t, kc, client)
-	testScaleUp(t, kc, client)
-	testScaleDown(t, kc, adminClient)
+	testScale(t, kc, client, data)
 
 	// cleanup
 	DeleteKubernetesResources(t, kc, testNamespace, data, templates)
-	cleanupServiceBusTopic(t, adminClient)
+	cleanupServiceBusTopic(t, adminClient, topicName)
 }
 
 func setupServiceBusTopicAndSubscription(t *testing.T) (*azservicebus.Client, *admin.Client) {
@@ -155,13 +161,33 @@ func setupServiceBusTopicAndSubscription(t *testing.T) (*azservicebus.Client, *a
 
 	_, err = adminClient.CreateTopic(context.Background(), topicName, nil)
 	assert.NoErrorf(t, err, "cannot create the topic - %s", err)
-	_, err = adminClient.CreateSubscription(context.Background(), topicName, subscriptionName, nil)
-	assert.NoErrorf(t, err, "cannot create the subscription - %s", err)
+
+	subscriptionName1 := fmt.Sprintf("%s-1", subscriptionPrefix)
+	setupSubscription(t, adminClient, topicName, subscriptionName1)
+	subscriptionName2 := fmt.Sprintf("%s-2", subscriptionPrefix)
+	setupSubscription(t, adminClient, topicName, subscriptionName2)
 
 	client, err := azservicebus.NewClientFromConnectionString(connectionString, nil)
 	assert.NoErrorf(t, err, "cannot connect to service bus namespace - %s", err)
 
 	return client, adminClient
+}
+
+func setupSubscription(t *testing.T, adminClient *admin.Client, topicName, subscriptionName string) {
+	_, err := adminClient.CreateSubscription(context.Background(), topicName, subscriptionName, nil)
+	assert.NoErrorf(t, err, "cannot create the subscription 1 - %s", err)
+
+	_, err = adminClient.DeleteRule(context.Background(), topicName, subscriptionName, "$Default", &admin.DeleteRuleOptions{})
+	assert.NoErrorf(t, err, "cannot delete default filter rule for subscription - %s", err)
+
+	ruleName := "filterRule"
+	_, err = adminClient.CreateRule(context.Background(), topicName, subscriptionName, &admin.CreateRuleOptions{
+		Name: &ruleName,
+		Filter: &admin.CorrelationFilter{
+			To: &subscriptionName,
+		},
+	})
+	assert.NoErrorf(t, err, "cannot create filter rule for subscription - %s", err)
 }
 
 func getTemplateData() (templateData, []Template) {
@@ -175,7 +201,8 @@ func getTemplateData() (templateData, []Template) {
 			TriggerAuthName:  triggerAuthName,
 			ScaledObjectName: scaledObjectName,
 			TopicName:        topicName,
-			SubscriptionName: subscriptionName,
+			SubscriptionName: fmt.Sprintf("%s.*", subscriptionPrefix),
+			Operation:        "sum",
 		}, []Template{
 			{Name: "secretTemplate", Config: secretTemplate},
 			{Name: "deploymentTemplate", Config: deploymentTemplate},
@@ -184,47 +211,42 @@ func getTemplateData() (templateData, []Template) {
 		}
 }
 
-func testActivation(t *testing.T, kc *kubernetes.Clientset, client *azservicebus.Client) {
-	t.Log("--- testing activation ---")
-	addMessages(t, client, 4)
-
-	AssertReplicaCountNotChangeDuringTimePeriod(t, kc, deploymentName, testNamespace, 0, 60)
-}
-
-func testScaleUp(t *testing.T, kc *kubernetes.Clientset, client *azservicebus.Client) {
+func testScale(t *testing.T, kc *kubernetes.Clientset, client *azservicebus.Client, data templateData) {
 	t.Log("--- testing scale up ---")
-	addMessages(t, client, 10)
 
-	assert.True(t, WaitForDeploymentReplicaReadyCount(t, kc, deploymentName, testNamespace, 1, 60, 1),
+	// send messages to subscription 1
+	addMessages(t, client, fmt.Sprintf("%s-1", subscriptionPrefix), 2)
+	// send messages to subscription 2
+	addMessages(t, client, fmt.Sprintf("%s-2", subscriptionPrefix), 4)
+
+	assert.True(t, WaitForDeploymentReplicaReadyCount(t, kc, deploymentName, testNamespace, 6, 60, 1),
 		"replica count should be 1 after 1 minute")
+
+	// check different aggregation operations
+	data.Operation = "max"
+	KubectlApplyWithTemplate(t, data, "scaledObjectTemplate", scaledObjectTemplate)
+	assert.True(t, WaitForDeploymentReplicaReadyCount(t, kc, deploymentName, testNamespace, 4, 60, 1),
+		"replica count should be 4 after 1 minute")
+
+	data.Operation = "avg"
+	KubectlApplyWithTemplate(t, data, "scaledObjectTemplate", scaledObjectTemplate)
+	assert.True(t, WaitForDeploymentReplicaReadyCount(t, kc, deploymentName, testNamespace, 3, 60, 1),
+		"replica count should be 3 after 1 minute")
 }
 
-func testScaleDown(t *testing.T, kc *kubernetes.Clientset, adminClient *admin.Client) {
-	t.Log("--- testing scale down ---")
-
-	_, err := adminClient.DeleteTopic(context.Background(), topicName, nil)
-	assert.NoErrorf(t, err, "cannot delete the topic - %s", err)
-	_, err = adminClient.CreateTopic(context.Background(), topicName, nil)
-	assert.NoErrorf(t, err, "cannot create the topic - %s", err)
-	_, err = adminClient.CreateSubscription(context.Background(), topicName, subscriptionName, nil)
-	assert.NoErrorf(t, err, "cannot create the subscription - %s", err)
-
-	assert.True(t, WaitForDeploymentReplicaReadyCount(t, kc, deploymentName, testNamespace, 0, 60, 1),
-		"replica count should be 0 after 1 minute")
-}
-
-func addMessages(t *testing.T, client *azservicebus.Client, count int) {
+func addMessages(t *testing.T, client *azservicebus.Client, subscriptionName string, count int) {
 	sender, err := client.NewSender(topicName, nil)
 	assert.NoErrorf(t, err, "cannot create the sender - %s", err)
 	for i := 0; i < count; i++ {
 		msg := fmt.Sprintf("Message - %d", i)
 		_ = sender.SendMessage(context.Background(), &azservicebus.Message{
 			Body: []byte(msg),
+			To:   &subscriptionName,
 		}, nil)
 	}
 }
 
-func cleanupServiceBusTopic(t *testing.T, adminClient *admin.Client) {
+func cleanupServiceBusTopic(t *testing.T, adminClient *admin.Client, topicName string) {
 	t.Log("--- cleaning up ---")
 	_, err := adminClient.DeleteTopic(context.Background(), topicName, nil)
 	assert.NoErrorf(t, err, "cannot delete service bus topic - %s", err)